requirements.txt
asgiref==3.5.2
Django==4.1.4
doc2text==0.2.4
docx2pdf==0.1.8
future==0.18.2
lxml==4.9.2
mime==0.1.0
numpy==1.24.0
packaging==22.0
Pillow==9.3.0
pycryptodome==3.16.0
pypdf2==2.12.1
pytesseract==0.3.10
python-docx==0.8.11
qrcode==7.3.1
reportlab==3.6.12
scipy==1.9.3
sqlparse==0.4.3
tqdm==4.64.1
typing-extensions==4.4.0
